ci: linux: Use a custom deploy script
Adds a bash re-implementation of linuxdeploy. Uses it instead of
linuxdeploy. The hope is to reduce copying a couple libraries that are
not required by yuzu.

In practice, this is only effective on machines that have unnecessary
libraries installed. linux-fresh is too stripped down to justify using
this script and the time penalty it incurs over linuxdeploy for zero
gains.